+++ src/internal/connector/exchange/api/mail.go
@@ -360,7 +360,7 @@ func (c Mail) Serialize(
 
 func MailInfo(msg models.Messageable) *details.ExchangeInfo {
 	var (
-		sender     = graph.UnwrapEmailAddress(msg.GetSender())
+		sender     = UnwrapEmailAddress(msg.GetSender())
 		subject    = ptr.Val(msg.GetSubject())
 		received   = ptr.Val(msg.GetReceivedDateTime())
 		created    = ptr.Val(msg.GetCreatedDateTime())
@@ -370,7 +370,7 @@ func MailInfo(msg models.Messageable) *details.ExchangeInfo {
 	if msg.GetToRecipients() != nil {
 		ppl := msg.GetToRecipients()
 		for _, entry := range ppl {
-			temp := graph.UnwrapEmailAddress(entry)
+			temp := UnwrapEmailAddress(entry)
 			if len(temp) > 0 {
 				recipients = append(recipients, temp)
 			}
@@ -386,4 +386,13 @@ func MailInfo(msg models.Messageable) *details.ExchangeInfo {
 		Created:   created,
 		Modified:  ptr.OrNow(msg.GetLastModifiedDateTime()),
 	}
+}
+
+func UnwrapEmailAddress(contact models.Recipientable) string {
+	var empty string
+	if contact == nil || contact.GetEmailAddress() == nil {
+		return empty
+	}
+
+	return ptr.Val(contact.GetEmailAddress().GetAddress())
 }
