--- src/internal/connector/graph/retry_middleware.go
@@ -1,114 +1,0 @@
-package graph
-
-import (
-	"context"
-	"net/http"
-	"strconv"
-	"time"
-
-	backoff "github.com/cenkalti/backoff/v4"
-	khttp "github.com/microsoft/kiota-http-go"
-)
-
-// ---------------------------------------------------------------------------
-// Client Middleware
-// ---------------------------------------------------------------------------
-
-// RetryHandler handles transient HTTP responses and retries the request given the retry options
-type RetryHandler struct {
-	// The maximum number of times a request can be retried
-	MaxRetries int
-	// The delay in seconds between retries
-	Delay time.Duration
-}
-
-func (middleware RetryHandler) retryRequest(
-	ctx context.Context,
-	pipeline khttp.Pipeline,
-	middlewareIndex int,
-	req *http.Request,
-	resp *http.Response,
-	executionCount int,
-	cumulativeDelay time.Duration,
-	exponentialBackoff *backoff.ExponentialBackOff,
-	respErr error,
-) (*http.Response, error) {
-	if (respErr != nil || middleware.isRetriableErrorCode(req, resp.StatusCode)) &&
-		middleware.isRetriableRequest(req) &&
-		executionCount < middleware.MaxRetries {
-		executionCount++
-
-		delay := middleware.getRetryDelay(req, resp, exponentialBackoff)
-
-		cumulativeDelay += delay
-
-		timer := time.NewTimer(delay)
-
-		select {
-		case <-ctx.Done():
-			// Don't retry if the context is marked as done, it will just error out
-			// when we attempt to send the retry anyway.
-			return resp, ctx.Err()
-
-		// Will exit switch-block so the remainder of the code doesn't need to be
-		// indented.
-		case <-timer.C:
-		}
-
-		req.Header.Set(retryAttemptHeader, strconv.Itoa(executionCount))
-
-		response, err := pipeline.Next(req, middlewareIndex)
-		if err != nil && !IsErrTimeout(err) {
-			return response, Stack(ctx, err).With("retry_count", executionCount)
-		}
-
-		return middleware.retryRequest(ctx,
-			pipeline,
-			middlewareIndex,
-			req,
-			response,
-			executionCount,
-			cumulativeDelay,
-			exponentialBackoff,
-			err)
-	}
-
-	if respErr != nil {
-		return nil, Stack(ctx, respErr).With("retry_count", executionCount)
-	}
-
-	return resp, nil
-}
-
-func (middleware RetryHandler) isRetriableErrorCode(req *http.Request, code int) bool {
-	return code == http.StatusInternalServerError || code == http.StatusServiceUnavailable
-}
-
-func (middleware RetryHandler) isRetriableRequest(req *http.Request) bool {
-	isBodiedMethod := req.Method == "POST" || req.Method == "PUT" || req.Method == "PATCH"
-	if isBodiedMethod && req.Body != nil {
-		return req.ContentLength != -1
-	}
-
-	return true
-}
-
-func (middleware RetryHandler) getRetryDelay(
-	req *http.Request,
-	resp *http.Response,
-	exponentialBackoff *backoff.ExponentialBackOff,
-) time.Duration {
-	var retryAfter string
-	if resp != nil {
-		retryAfter = resp.Header.Get(retryAfterHeader)
-	}
-
-	if retryAfter != "" {
-		retryAfterDelay, err := strconv.ParseFloat(retryAfter, 64)
-		if err == nil {
-			return time.Duration(retryAfterDelay) * time.Second
-		}
-	} // TODO parse the header if it's a date
-
-	return exponentialBackoff.NextBackOff()
-}
